agents/requirements.txt
browser-use==0.12.0
httpx>=0.27.2,<1.0.0
yt-dlp>=2024.1.0
python-dotenv>=1.0.0
flask>=3.0.0
